fix cgo

+++ kwdbts2/engine/db.cpp
@@ -1179,8 +1179,8 @@ TSStatus TSDeleteSnapshot(TSEngine* engine, TSTableID table_id, uint64_t snapsho
   return kTsSuccess;
 }
 
-TSStatus TSReadBatchData(TSEngine* engine, TSTableID table_id, uint32_t table_version, uint64_t begin_hash,
-                         uint64_t end_hash, KwTsSpan ts_span, uint64_t job_id, TSSlice* data, int32_t* row_num) {
+TSStatus TSReadBatchData(TSEngine* engine, TSTableID table_id, uint64_t table_version, uint64_t begin_hash,
+                         uint64_t end_hash, KwTsSpan ts_span, uint64_t job_id, TSSlice* data, uint32_t* row_num) {
   kwdbContext_t context;
   kwdbContext_p ctx = &context;
   KStatus s = InitServerKWDBContext(ctx);
@@ -1195,7 +1195,7 @@ TSStatus TSReadBatchData(TSEngine* engine, TSTableID table_id, uint32_t table_ve
 }
 
 TSStatus TSWriteBatchData(TSEngine* engine, TSTableID table_id, uint64_t table_version, uint64_t job_id,
-                          TSSlice* data, int32_t* row_num) {
+                          TSSlice* data, uint32_t* row_num) {
   kwdbContext_t context;
   kwdbContext_p ctx = &context;
   KStatus s = InitServerKWDBContext(ctx);
