+++ kwdbts2/engine/include/engine.h
@@ -351,14 +351,14 @@ struct TSEngine {
     return KStatus::FAIL;
   }
 
-  virtual KStatus ReadBatchData(kwdbContext_p ctx, TSTableID table_id, uint32_t table_version, uint64_t begin_hash,
+  virtual KStatus ReadBatchData(kwdbContext_p ctx, TSTableID table_id, uint64_t table_version, uint64_t begin_hash,
                                 uint64_t end_hash, KwTsSpan ts_span, uint64_t job_id, TSSlice* data,
-                                int32_t* row_num) {
+                                uint32_t* row_num) {
     return FAIL;
   }
 
   virtual KStatus WriteBatchData(kwdbContext_p ctx, TSTableID table_id, uint64_t table_version, uint64_t job_id,
-                                 TSSlice* data, int32_t* row_num) {
+                                 TSSlice* data, uint32_t* row_num) {
     return FAIL;
   }
